requirements.txt
Flask==1.0.2
gunicorn==19.9.0
pyrebase4
